Plugin [Completed]
This was my associate capstone project; Plugin.
It features more advanced CRUD functionality for users to send message, confirm sales/purchases. It is akin to a lite ebay.

MUST be hosted on XAMPP, and the db was hosted with my university's servers. This is archived simply for the logic of the program.